add adapter class for string offsets
